Unit test client complete

+++ src/modbus.rs
@@ -174,6 +174,7 @@ impl Default for Timeout {
 /// The different parts of libmodbus are implemented as traits. The modules of this crate contains these
 /// traits and a implementation with a, hopefully safe, interface.
 ///
+#[derive(Debug)]
 pub struct Modbus {
     pub ctx: *mut ffi::modbus_t,
 }
@@ -192,7 +193,7 @@ impl Modbus {
     pub const RTU_MAX_ADU_LENGTH: usize = ffi::MODBUS_RTU_MAX_ADU_LENGTH as usize;
     pub const TCP_DEFAULT_PORT: u32 = ffi::MODBUS_TCP_DEFAULT_PORT;
     pub const TCP_MAX_ADU_LENGTH: usize = ffi::MODBUS_TCP_MAX_ADU_LENGTH as usize;
-    pub const TCP_SLAVE: u32 = ffi::MODBUS_TCP_SLAVE;
+    pub const TCP_SLAVE: u8 = ffi::MODBUS_TCP_SLAVE as u8;
     pub const BROADCAST_ADDRESS: u8 = ffi::MODBUS_BROADCAST_ADDRESS as u8;
 
     /// `connect` - establish a Modbus connection
